`filter` now accepts key, not items and index.
The optional `filter` function now accepts the key for an item instead
of the array of items and the index. We never meant for the user to
filter based on the body of a record, just the key.

Now we want to be able to select multiple results for a particular key,
in order to produce something meaningful when someone uses this `map`
function against an index in Memento so we're going to have to look at
the right key and possibly continue with the next page.

Closes #20.