Continue partials across multiple pages.
Used to assume that all of the entries for a partial would be on the
same page since this was written with Memento's `map` in mind, but a map
of a non-unique index, can return zero, one, or more results. Was
considering only mapping items out of a table and not an index and the
tables are required to have a unique key in Memento so that was zero or
one.

Here check to see if the filter function matches the right key if we are
at the end of the items array and if so we continue trying to match with
the filter on the next page.

Closes #21.