Return extracted key.
The extracted key is now returned with the value in a structure called
`sought` in the array of results. The `key` is now the key of the item
or `null` if the item was not found. This gives the user everything that
Skip had to work with. By default, it is not in the correct shape for
use with the rest of the MVCC library since we won't know how to create
a meaningful `null` key from a partial key. This is up to the user and
it can be done with `nullify` which was originally intended for use to
create a delete entry for Homogenize, but it can be used for keys as
well.

Closes #25.