Group sets by entry.
Wasn't working with Homogenize which expects the records sorted either
ascending or descending. With partial a set ordered `[ 'b', 'a' ]` the
key `[ 'b' ]` comes before `[ 'a', 1 ]` and `[ 'b', 1 ]` both, it's up
next, but on the next selection `[ 'a', 1 ]` comes before `[ 'b', 1 ]`
so now we are out of order.

For sets we're creating groups and we'll have to have a special `map`
for Homogenize that goes through the sets item by item merging the
`item` list in each set.

Record is now simpler and easier to understand. Index is implicit and
can be presented by a transformation that flattens into the old format,
or maybe this is the end format for applications like Memento.

Closes #28.